pyproject.toml
[project]
name = "genai_evaluator"
version = "0.1.0"
description = "Add your description here"
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "azure-core>=1.34.0",
    "azure-identity>=1.22.0",
    "jupyter>=1.1.1",
    "levenshtein>=0.27.1",
    "nltk>=3.9.1",
    "numpy>=2.1.3",
    "openai>=1.57.0",
    "promptflow>=1.18.0",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.1.0",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=4.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "ruff>=0.8.2",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Import order
    "I",
]
ignore = [
    #exclude line too long warning
    "E501",
    "D104"
]
